+++ src/vector3d.cpp
@@ -2,11 +2,11 @@
 
 
 cml::Vector3::Vector3(): x(0.0f), y(0.0f), z(0.0f){}
-cml::Vector3::Vector3(double uX, double uY, double uZ): x(uX), y(uY), z(uZ){}
+cml::Vector3::Vector3(float uX, float uY, float uZ): x(uX), y(uY), z(uZ){}
 cml::Vector3::Vector3(const Vector3& v): x(v.x), y(v.y), z(v.z){}
 cml::Vector3::~Vector3(){}
 
-cml::Vector3 cml::scalar(double sc, Vector3 &v){
+cml::Vector3 cml::scalar(float sc, Vector3 &v){
     return Vector3(sc*v.x, sc*v.y, sc*v.z);
 }
 
@@ -14,20 +14,20 @@ cml::Vector3 cml::cross(Vector3 &a, Vector3 &b){
     return Vector3(a.y*b.z - a.z*b.y, a.z*b.x - a.x*b.z, a.x*b.y - a.y*b.x);
 }
 
-double cml::dot(Vector3 &a, Vector3 &b){
+float cml::dot(Vector3 &a, Vector3 &b){
     return a.x*b.x + a.y*b.y + a.z*b.z;
 }
 
-double cml::magnitude(Vector3 &v){
+float cml::magnitude(Vector3 &v){
     return sqrt(pow(v.x, 2) + pow(v.y, 2) + pow(v.z, 2));
 }
 
-double cml::angle(Vector3 &a, Vector3 &b){
+float cml::angle(Vector3 &a, Vector3 &b){
     return acos(dot(a, b)/magnitude(a) * magnitude(b));
 }
 
 cml::Vector3 cml::unit(Vector3 &v){
-    double vMagnitude = magnitude(v);
+    float vMagnitude = magnitude(v);
     return Vector3(v.x/vMagnitude, v.y/vMagnitude, v.z/vMagnitude);
 }
 
@@ -36,7 +36,7 @@ cml::Vector3 cml::projection(Vector3 &a, Vector3 &b){
     return result;
 }
 
-double cml::scalar_projection(Vector3 &a, Vector3 &b){
+float cml::scalar_projection(Vector3 &a, Vector3 &b){
     return dot(b, a) / magnitude(a);
 }
 
@@ -69,11 +69,11 @@ cml::Vector3 cml::Vector3::operator*(const float s)const{
     return Vector3(this->x*s, this->y*s, this->z*s);
 }
 
-double cml::Vector3::operator*(const Vector3& v)const{
+float cml::Vector3::operator*(const Vector3& v)const{
     return x*v.x+y*v.y+z*v.z;
 }
 
-double cml::Vector3::dot(const Vector3& v) const{
+float cml::Vector3::dot(const Vector3& v) const{
     return x*v.x+y*v.y+z*v.z;
 }
